+++ pkg/cmd/pulumi/project/project.go
@@ -37,6 +37,7 @@ func NewProjectCmd() *cobra.Command {
 	// Add subcommands
 	cmd.AddCommand(newProjectLsCmd())
 	cmd.AddCommand(newcmd.NewNewCmd())
+	cmd.AddCommand(newInitCmd())
 
 	return cmd
 }
