+++ pkg/cmd/pulumi/pulumi.go
@@ -60,7 +60,6 @@ import (
 	cmdDo "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/do"
 	cmdEnv "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/env"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/events"
-	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/initcmd"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/insights"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/install"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/logs"
@@ -450,7 +449,6 @@ func NewPulumiCmd() (*cobra.Command, func()) {
 		{
 			Name: "Stack Management Commands",
 			Commands: []*cobra.Command{
-				initcmd.NewInitCmd(),
 				newcmd.NewNewCmd(),
 				config.NewConfigCmd(pkgWorkspace.Instance),
 				cmdStack.NewStackCmd(),
